StoredEventRepository should not be opinionated
The StoredEventRepository interface should not impose an int $startingFrom type as another repository implementation may use a different lookup method.

An example would be for a repository which does not use an auto-incrementing primary key and, instead of returning the events ordered by ID, returns the events ordered by milliseconds, microseconds or nanoseconds. In this implementation retrieveAllStartingFrom may expect a float, string, or some other variable type to appropriately handle the offset and replay the events.